Aligned coded.
