+++ blueprints/c-component.vue
@@ -44,7 +44,7 @@
 </script>
 
 <style lang="scss">
-  // @use '../setup/scss/variables';
+  // @use '@/setup/scss/variables';
 
   .c-component {
     // TODO: update selector according to filename
