+++ src/components/c-header.vue
@@ -5,13 +5,19 @@
 <script lang="ts">
   import { defineComponent } from 'vue';
 
+  // type Setup = {};
+  // type Data = {};
+
   export default defineComponent({
     name: 'c-header',
 
     // components: {},
 
     // props: {},
-    // data() {
+    // setup(): Setup {
+    //   return {};
+    // },
+    // data(): Data {
     //   return {};
     // },
