+++ src/elements/e-button.vue
@@ -32,6 +32,8 @@
   import { Modifiers } from '@/plugins/vue-bem-cn/src/globals';
   import eProgress from './e-progress.vue';
 
+  // type Setup = {};
+
   type Attributes = {
     role: string | null;
     disabled: boolean;
@@ -155,6 +157,9 @@
       },
     },
 
+    // setup(): Setup {
+    //   return {};
+    // },
     data(): Data {
       return {
         /**
