+++ src/elements/e-multiselect.vue
@@ -272,17 +272,13 @@
 
       /**
        * Shows if the disabled state of the component should be active.
-       *
-       * @returns {boolean}
        */
-      isDisabled() {
+      isDisabled(): boolean {
         return this.disabled || this.progress;
       },
 
       /**
        * Gets the filtered options if the user used the search.
-       *
-       * @returns {array.<object>}
        */
       filteredOptions(): Option[] {
         if (this.hasSearch && this.searchTerm) {
