+++ src/elements/e-select.vue
@@ -51,6 +51,8 @@
   import eProgress from '@/elements/e-progress.vue';
   import { Modifiers } from '@/plugins/vue-bem-cn/src/globals';
 
+  // type Setup = {};
+
   type Data = {
     internalValue: string;
   };
