+++ src/plugins/tooltip/styles/styles.scss
@@ -1,16 +1,15 @@
-@use '../../../setup/scss/variables' as *;
-@use '../../../setup/scss/mixins' as *;
-@use '../../../setup/scss/functions' as *;
+@use '../../../setup/scss/variables';
+@use '../../../setup/scss/mixins';
 
 .c-tooltip {
   $this: &;
-  $background-color: $color-grayscale--500;
-  $color: $color-grayscale--0;
-  $box-shadow-color: rgba($color-grayscale--0, 0.2);
-  $border-radius: $border-radius--500;
+  $background-color: variables.$color-grayscale--500;
+  $color: variables.$color-grayscale--0;
+  $box-shadow-color: rgba(variables.$color-grayscale--0, 0.2);
+  $border-radius: variables.$border-radius--500;
   $pointer__width: 6px;
   $wrapper__padding: $pointer__width + 6px;
-  $content__padding: $spacing--5 $spacing--10;
+  $content__padding: variables.$spacing--5 variables.$spacing--10;
 
   position: relative;
   display: block;
@@ -32,10 +31,10 @@
     max-width: 80vw;
     opacity: 0;
     color: $color;
-    transition: opacity $transition-duration--100;
+    transition: opacity variables.$transition-duration--100;
     pointer-events: none;
 
-    @include media(sm) {
+    @include mixins.media(sm) {
       max-width: 400px;
     }
 
@@ -88,8 +87,8 @@
       width: 0;
       height: 0;
       border: $pointer__width solid transparent;
-      border-bottom-color: $color-grayscale--500;
-      border-left-color: $color-grayscale--500;
+      border-bottom-color: variables.$color-grayscale--500;
+      border-left-color: variables.$color-grayscale--500;
       background: $background-color;
 
       // Left center
@@ -116,7 +115,7 @@
       // Top start
       [data-popper-placement='top-start'] &,
       [data-popper-placement='bottom-start'] & {
-        left: $wrapper__padding + $spacing--10;
+        left: $wrapper__padding + variables.$spacing--10;
       }
 
       // Top center
@@ -128,7 +127,7 @@
     }
 
     &::after {
-      @include z-index(back);
+      @include mixins.z-index(back);
 
       box-shadow: 0 2px 6px $box-shadow-color;
     }
@@ -137,7 +136,7 @@
   &__transition {
     &--fade-enter-active,
     &--fade-leave-active {
-      transition: opacity $transition-duration--200 ease;
+      transition: opacity variables.$transition-duration--200 ease;
     }
 
     &--fade-enter-from,
