+++ src/styleguide/components/s-api-mock-test.vue
@@ -314,8 +314,8 @@
 </script>
 
 <style lang="scss">
-  @use '../../setup/scss/variables';
-  @use '../../setup/scss/mixins';
+  @use '@/setup/scss/variables';
+  @use '@/setup/scss/mixins';
 
   .s-api-mock-test {
     @include mixins.z-index(front);
