+++ src/styleguide/demo-pages/sandbox/r-forms.vue
@@ -418,7 +418,7 @@
 </script>
 
 <style lang="scss">
-  @use '../../../setup/scss/variables';
+  @use '@/setup/scss/variables';
 
   .r-forms {
     display: grid;
