+++ src/styleguide/demo-pages/sandbox/r-wysiwyg.vue
@@ -106,13 +106,21 @@
 <script lang="ts">
   import { defineComponent } from 'vue';
 
+  // type Setup = {};
+  // type Data = {};
+
   export default defineComponent({
     name: 'r-wysiwyg',
 
     // components: {},
 
     // props: {},
-    // data() {
+    // emits: {},
+
+    // setup(): Setup {
+    //   return {};
+    // },
+    // data(): Data {
     //   return {};
     // },
